+++ modules/flipcache/example/json.js
@@ -1,0 +1,6 @@
+const log = require('fliplog')
+const flipcache = require('../src')
+
+// @TODO: config
+// const datas = flipcache.from('./index.js').load().json().update({eh: true}).parse().clean()
+// log.data(datas).bold('flip').echo()