+++ modules/flipcache/test/json.js
@@ -1,4 +1,0 @@
-const log = require('fliplog')
-
-const datas = flipcache.from('./index.js').load().json().update({eh: true}).parse().clean()
-log.data(datas).bold('flip').echo()