+++ modules/flipcache/test/restoring.js
@@ -1,26 +1,16 @@
-// -----------
-
+const test = require('ava')
 const flipcache = require('../src')
 
-const mocha = flipcache
-  .from('mochaOptsFile')
-  .to('mochaOptsBackupFile')
-  .autoRestore(2000)
-
-// is a .to example
-// on-the-fly
-const onthefly = flipcache
-  .from('fromfile')
-  .to('otherfile')
-  .autoRemove(2000)
+test.todo('config')
+test.todo('restore')
+test.todo('from to')
 
 const config = flipcache
   .to('.config.js')
   .json()
   .load()
-  .key('eh', ['some values'])
+  .update('eh', ['some values'])
   .write()
-
 
 // is a .to example
 // on-the-fly